Made the GlobalModel data access methods throwable
it gives more opportunities to use it, turn throwable function result into optional result is as easy as use ‘try?’ in front of the function call.
NOTE: that operators are still going to return optionals, to avoid confusions. Operators act like subscript, as it’s hard to discover that operator throws.