Connected ‘Scenario’ with its ‘Story’
it’s not a type/protocol with generic type to difine the Story type, becasue it’s not supposed to be declared as a type in one place and then used in another place. It’s a convenience business logic container that helps to define a workflow in place, without maintaining reference to it in any way.