[SIMPLIFY] initialization of `ModelContainer`
Use DEDICATED struct as ViewModel, it will have @Published properties that we can set directly for debug/preview, but in runtime we can also use assign(to:) from publishers from dispatcher. Because it's view model - all props must have default values and we can init them using ?= and try? to init with values from dispatcher
in runtime, but it won't fail/interfer in debug/preview.